refactor: removed 'logic' folder and moved 'ContactPage' to components folder as it doesn't hold any logic rather UI stuff
